feat: rework search as a component
